auth_service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["Ophely-git <ophely1992@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.8"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
jwt = "^1.3.1"
google-auth = "^2.32.0"
celery = "^5.4.0"
requests = "^2.32.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
